+++ clang-tools-extra/clang-tidy/kokkos/KokkosMatchers.cpp
@@ -16,17 +16,54 @@ namespace clang {
 namespace tidy {
 namespace kokkos {
 
-bool explicitlyUsingHostExecutionSpace(CallExpr const *CE,
-                                       std::string const &RegexString) {
+namespace {
+TypedefNameDecl const *getTypedefFromFirstTemplateArg(Expr const *E) {
+  if (E == nullptr) {
+    return nullptr;
+  }
+
+  auto const *TST = E->getType()->getAs<TemplateSpecializationType>();
+  if (TST == nullptr) {
+    return nullptr;
+  }
+  if (TST->getNumArgs() < 1) {
+    return nullptr;
+  }
+
+  auto const *TDT = TST->getArg(0).getAsType()->getAs<TypedefType>();
+  if (TDT == nullptr) {
+    return nullptr;
+  }
+
+  auto const *TDD = dyn_cast_or_null<TypedefNameDecl>(TDT->getDecl());
+  return TDD;
+}
+} // namespace
+
+bool matchesAnnotation(Decl const *D, std::string const &RegExp) {
+  assert(!RegExp.empty());
+  llvm::Regex Re(RegExp);
+  for (auto const *Attr : D->attrs()) {
+    if (auto const *Anna = dyn_cast<AnnotateAttr>(Attr)) {
+      if (Re.match(Anna->getAnnotation())) {
+        return true;
+      }
+    }
+  }
+  return false;
+}
+
+
+bool explicitlyDefaultHostExecutionSpace(CallExpr const *CE) {
   using namespace clang::ast_matchers;
   auto &Ctx = CE->getCalleeDecl()->getASTContext();
 
   // We will assume that any policy where the user might explicitly ask for the
   // host space inherits from Impl::PolicyTraits
-  auto FilterArgs =
-      hasAnyArgument(expr(hasType(cxxRecordDecl(isDerivedFrom(cxxRecordDecl(
-                              matchesName("Impl::PolicyTraits"))))))
-                         .bind("expr"));
+  auto FilterArgs = hasAnyArgument(
+      expr(hasType(classTemplateSpecializationDecl(isDerivedFrom(
+               cxxRecordDecl(matchesName("Impl::PolicyTraits"))))))
+          .bind("expr"));
 
   // We have to jump through some hoops to find this, if we just looked at the
   // template type of the Policy constructor we lose the sugar and instead of
@@ -34,15 +71,11 @@ bool explicitlyUsingHostExecutionSpace(CallExpr const *CE,
   // to such as Kokkos::Serial, preventing us from figuring out if the user
   // actually asked for a host space specifically or just happens to have a
   // host space as the default space.
-  llvm::Regex Reg(RegexString);
-  auto BNs = match(callExpr(FilterArgs).bind("CE"), *CE, Ctx);
+  auto BNs = match(callExpr(FilterArgs), *CE, Ctx);
   for (auto &BN : BNs) {
-    if (auto const *E = BN.getNodeAs<Expr>("expr")) {
-      if (auto const *TST = E->getType()->getAs<TemplateSpecializationType>()) {
-        if (Reg.match(TST->getArg(0).getAsType().getAsString())) {
-          return true;
-        }
-      }
+    if (auto const *TDD =
+            getTypedefFromFirstTemplateArg(BN.getNodeAs<Expr>("expr"))) {
+      return matchesAnnotation(TDD, "DefaultHostExecutionSpace");
     }
   }